+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - FIX: Fixing stack when saved as integer for some special cases 
 - FIX: Clipping negative reflectances to 0([#79](https://github.com/sertit/eoreader/issues/79)
+- FIX: Fixing nodata mangement for Theia product
 
 ### Other